requirements.txt
asyncpraw==7.3.0
discord.py[voice]==1.7.3
packaging>=20.9
python-dotenv==0.18.0
streamlink==2.3.0
# New in version 1.1.0
yt-dlp==2021.9.2
# New in version 2.0.0
asyncpg==0.24.0
# Deprecated since version 1.1.0
#youtube-dl==2021.6.6
